- Added c_runtime/fortran_types.h   + Contains only fortran_integer - the same size integer as the ones used by the DASSL solver. - Updated solver_dasrt.h to use fortran_integer instead of long.   + Fixes some of the new 64-bit issues.   + Also updated functions that were used by this function.   + Updated code generation for the new function defintions. - The testsuite now runs on 64-bit machines (except some MetaModelica tests)
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@4918 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ c_runtime/index_spec.c
@@ -120,7 +120,7 @@ void print_size_array(int size, size_t* arr)
 	int i;
 	printf("{");
 	for(i = 0; i < size; ++i) {
-		printf("%d", arr[i]);
+		printf("%d", (int) arr[i]);
 		if (i != (size - 1)) printf(",");
 	}
 	printf("}\n");
